rptest: Move tests to separate folder
