+++ tests/rptest/scale_tests/many_partitions_test.py
@@ -33,8 +33,6 @@ from rptest.services.kgo_verifier_services import KgoVerifierProducer, KgoVerifi
 from rptest.services.kgo_repeater_service import KgoRepeaterService, repeater_traffic
 from rptest.services.openmessaging_benchmark import OpenMessagingBenchmark
 from rptest.services.openmessaging_benchmark_configs import OMBSampleConfigurations
-from rptest.services.producer_swarm import ProducerSwarm
-from rptest.services.consumer_swarm import ConsumerSwarm
 from rptest.utils.scale_parameters import ScaleParameters
 from rptest.util import inject_remote_script
 
@@ -999,626 +997,3 @@ class ManyPartitionsTest(PreallocNodesTest):
                     f"Expected throughput {expect_mbps:.2f}MiB/s, got throughput {actual_mbps:.2f}MiB/s"
                 )
                 raise
-
-    def _try_parse_json(self, node, jsondata):
-        try:
-            return json.loads(jsondata)
-        except ValueError:
-            self.logger.debug(
-                f"{str(node.account)}: Could not parse as json: {str(jsondata)}"
-            )
-            return None
-
-    def _produce_messages_to_random_topics(self, kclient, message_count,
-                                           num_topics, topic_names):
-        """Select random num_topics from the list topic_names
-        and send message_count to it with consecutive numbers as values
-
-        Args:
-            kclient (_type_): python_librdkafka client
-            message_count (_type_): number of messages to produce
-            num_topics (_type_): number of random topics to produce to
-            topic_names (_type_): list of topic names
-
-        return values:
-            used_topics_list, ununsed_topics_list, errors
-        """
-        def _send_messages(topic):
-            """
-                Simple function that sends indices as messages
-            """
-            errors = []
-
-            def acked(err, msg):
-                """
-                    Simple and unsafe callback
-                """
-                if err is not None:
-                    errors.append(f"FAIL: {str(msg)}: {str(err)}")
-
-            # Sent indices as values
-            p = kclient.get_producer()
-            sent_count = 0
-            for idx in range(1, message_count + 1):
-                # Async message sending func
-                p.produce(topic,
-                          key=f"key_{idx}",
-                          value=f"{idx:04}",
-                          callback=acked)
-                # Make sure message sent, aka sync
-                p.flush()
-                sent_count += 1
-            # Return stats and errors
-            return sent_count, errors
-
-        # Pick random topics to send messages to
-        total_topics = len(topic_names) - 1
-        if total_topics > num_topics:
-            random_topic_indices = [
-                random.randint(0, total_topics) for i in range(num_topics)
-            ]
-            next_topic_batch = []
-            while len(random_topic_indices) > 0:
-                next_topic_batch.append(
-                    topic_names[random_topic_indices.pop()])
-            new_topic_names = list(set(topic_names) - set(next_topic_batch))
-        else:
-            next_topic_batch = topic_names
-            new_topic_names = []
-
-        # Send messages
-        messages_sent = 0
-        errors = []
-        with concurrent.futures.ThreadPoolExecutor(16) as executor:
-            for c, thread_errors in executor.map(_send_messages,
-                                                 next_topic_batch):
-                messages_sent += c
-                errors += thread_errors
-        self.logger.info(f"Total of {messages_sent} messages sent")
-
-        total_errors = len(errors)
-        if total_errors > 0:
-            self.logger.error(f"{total_errors} Errors detected "
-                              "while sending messages")
-
-        return next_topic_batch, new_topic_names, errors
-
-    def _consume_messages_from_random_topic(self,
-                                            kclient,
-                                            message_count,
-                                            topic_names,
-                                            timeout_sec=300):
-        """Consume message_count from random topic in the list topic_names
-
-        Args:
-            kclient (_type_): python_librdkafka client
-            message_count (_type_): number of messages to consume
-            topic_names (_type_): list of topic names to select from
-            timeout_sec (int, optional): Timeout. Defaults to 300.
-
-        Raises:
-            KafkaException: On Kafka transport errors
-            RuntimeError: On timeout consuming messages
-            RuntimeError: On non-consecutive values in messages
-
-        Returns: None
-        """
-
-        # Function checks if numbers in list are consecutive
-        def check_consecutive(numbers_list):
-            n = len(numbers_list) - 1
-            # Calculate iterative difference for the array. It should be 1.
-            return (sum(numpy.diff(sorted(numbers_list)) == 1) >= n)
-
-        # Select random topic from the list
-        target_topic = topic_names[random.randint(0, len(topic_names))]
-        # Consumer specific config
-        consumer_extra_config = {
-            "auto.offset.reset": "smallest",
-            "group.id": "topic_swarm_group"
-        }
-        self.logger.info(f"Start consuming from {target_topic}")
-        # Consumer
-        start_time_s = time.time()
-        consumer = kclient.get_consumer(consumer_extra_config)
-        numbers = []
-        # Message consuming loop
-        try:
-            consumer.subscribe([target_topic])
-            while True:
-                # Poll for the message
-                msg = consumer.poll(timeout=1.0)
-                if msg is None:
-                    continue
-                # On error, check for the EOF
-                if msg.error():
-                    if msg.error().code() == KafkaError._PARTITION_EOF:
-                        # End of partition event
-                        self.logger.info(f"Consumer of '{msg.topic()}' "
-                                         f"[{msg.partition()}] reached "
-                                         f"end at offset {msg.offset()}")
-                        break
-                    # If not EOF, raise it
-                    elif msg.error():
-                        raise KafkaException(msg.error())
-                else:
-                    # Save value from the message
-                    numbers.append(int(msg.value()))
-                # calculate elapsed time
-                elapsed = time.time() - start_time_s
-                # Exit on target number reached or timeout
-                if len(numbers) == message_count or elapsed > timeout_sec:
-                    break
-        finally:
-            # Close down consumer to commit final offsets.
-            consumer.close()
-
-        self.logger.info(f"Consumed {len(numbers)} messages")
-        # Check that we received all numbers
-        if elapsed > timeout_sec:
-            raise RuntimeError("Timeout consuming messages "
-                               f"from {target_topic}")
-        elif not check_consecutive(numbers):
-            raise RuntimeError("Produced and consumed messages mismatch "
-                               f"for {target_topic}")
-
-        return
-
-    def _write_and_random_read_many_topics(self, message_count, num_topics,
-                                           topic_names):
-        """
-            Test checks that each of the many topics can be written to.
-            This produce/consume implementation will check actual data
-            of the messages to ensure that all of the messages are delivered
-
-            Pick X number of topics, write 100 messages in each
-            Pick random one among them, consume all messages
-            Iterate.
-        """
-        # Prepare librdkafka python client
-        kclient = PythonLibrdkafka(self.redpanda)
-        self.logger.info(
-            f"Producing {message_count} messages to {num_topics} random topics"
-        )
-
-        # Produce messages
-        used_topics, unused_topics, errors = \
-            self._produce_messages_to_random_topics(kclient, message_count,
-                                                    num_topics, topic_names)
-
-        # Consume messages
-        # Will raise RuntimeException on timeout
-        # or non-consecutive message values
-        self._consume_messages_from_random_topic(kclient,
-                                                 message_count,
-                                                 topic_names,
-                                                 timeout_sec=300)
-
-        # Return list of topics that was not used
-        return unused_topics, errors
-
-    def _create_many_topics(self,
-                            brokers,
-                            node,
-                            topic_name_prefix,
-                            topic_count,
-                            batch_size,
-                            num_partitions,
-                            num_replicas,
-                            use_kafka_batching,
-                            topic_name_length=200,
-                            skip_name_randomization=False):
-        """Function uses batched topic creation approach.
-        Its either single topic per request using ThreadPool in batches or
-        the whole batch in single kafka request.
-
-        Args:
-            topic_count (int): Number of topics to create
-            batch_size (int): Batch size for one create operation
-            topic_name_length (int): Total topic length for randomization
-            num_partitions (int): Number of partitions per topic
-            num_replicas (int): Number of replicas per topic
-            use_kafka_batching (bool): on True sends whole batch as a single
-            request.
-
-        Raises:
-            RuntimeError: Underlying creation script generated error
-
-        Returns:
-            list: topic names and timing data
-        """
-        def log_timings_with_percentiles(timings):
-            # Extract data
-            created_count = timings.get('count_created', -1)
-
-            # Add min/max time to the log
-            tmin = timings.get('creation-time-min', 0)
-            tmax = timings.get('creation-time-max', 0)
-            tp_str = f"...{created_count} topics: " \
-                "min = {:>7,.3f}s, " \
-                "max = {:>7,.3f}s, ".format(tmin, tmax)
-
-            # Calculate percentiles for latest batch
-            prc = [25, 50, 75, 90, 95, 99]
-            creation_times = timings.get('creation_times', [])
-            tprc = numpy.percentile(creation_times, prc)
-            for i in range(len(prc)):
-                tp_str += "p{} ={:7,.3f}s, ".format(prc[i], tprc[i])
-            # Log them
-            self.logger.debug(tp_str)
-
-        # Prepare command
-        remote_script_path = inject_remote_script(node, "topic_operations.py")
-        cmd = f"python3 {remote_script_path} "
-        cmd += f"--brokers '{brokers}' "
-        cmd += f"--batch-size '{batch_size}' "
-        cmd += "create "
-        cmd += f"--topic-prefix '{topic_name_prefix}' "
-        cmd += f"--topic-count {topic_count} "
-        cmd += "--kafka-batching " if use_kafka_batching else ""
-        cmd += f"--topic-name-length {topic_name_length} "
-        cmd += f"--partitions {num_partitions} "
-        cmd += f"--replicas {num_replicas} "
-        cmd += "--skip-randomize-names" if skip_name_randomization else ""
-        hostname = node.account.hostname
-        self.logger.info(f"Starting topic creation script on '{hostname}")
-        self.logger.debug(f"...cmd: {cmd}")
-
-        data = {}
-        for line in node.account.ssh_capture(cmd):
-            # The script will produce jsons one, per line
-            # And it will have timings only
-            self.logger.debug(f"received {sys.getsizeof(line)}B "
-                              f"from '{hostname}'.")
-            data = self._try_parse_json(node, line.strip())
-            if data is not None:
-                if 'error' in data:
-                    self.logger.warning(f"Node '{hostname}' reported "
-                                        f"error:\n{data['error']}")
-                    raise RuntimeError(data['error'])
-                else:
-                    # Extract data
-                    timings = data.get('timings', {})
-                    log_timings_with_percentiles(timings)
-            else:
-                data = {}
-
-        topic_details = data.get('topics', [])
-        current_count = len(topic_details)
-        self.logger.info(f"Created {current_count} topics")
-        assert len(topic_details) == topic_count, \
-            f"Topic count not reached: {current_count}/{topic_count}"
-
-        return topic_details
-
-    @cluster(num_nodes=10, log_allow_list=RESTART_LOG_ALLOW_LIST)
-    def test_topic_swarm(self):
-        """Test creates 11950 topics, validates partitions and replicas,
-        produces 100 messages to each topic using batches, consumes
-        messages from random topic in each batch and validates message content
-
-        Returns:
-            None
-        """
-
-        # Max number of partitions for i3en.xlarge
-        # default settings is 1000 partitions per shard/cpu
-        # (9 nodes × 4 vcpus/shards × 1000) / 3 replicas
-        # 12000
-        tsm = TopicScaleProfileManager()
-        profile = tsm.get_profile("topic_profile_t10k_p1")
-
-        # Start kafka
-        self.redpanda.start()
-
-        # Brokers list suitable for script arguments
-        brokers = ",".join(self.redpanda.brokers_list())
-
-        # With current settings, there should be single available node
-        node = self.cluster.alloc(ClusterSpec.simple_linux(1))[0]
-
-        # Call function to create the topics
-        topic_details = self._create_many_topics(
-            brokers,
-            node,
-            profile.topic_name_prefix,
-            profile.topic_count,
-            profile.batch_size,
-            profile.num_partitions,
-            profile.num_replicas,
-            profile.use_kafka_batching,
-            topic_name_length=profile.topic_name_length,
-            skip_name_randomization=False)
-
-        # Validate topics
-        topics = self.rpk.list_topics(detailed=True)
-
-        # Validate topic creation
-        self.logger.info("Validating created topics")
-        topics_ok = []
-        topics_failed = []
-
-        # Searches for a topic in our list
-        def get_topic_index(name):
-            for idx in range(len(topic_details)):
-                if name == topic_details[idx]['name']:
-                    return idx
-            return -1
-
-        # Validate RP topics against requested details
-        for name, partitions, replicas in topics:
-            err = "ok"
-            idx = get_topic_index(name)
-            if idx < 0:
-                # RP has topic that is not ours
-                # Cluster was clean and it is an issue
-                err = f"Unexpected topic found: {name}"
-            elif int(replicas) != topic_details[idx]['replicas']:
-                # Replication factor is wrong
-                err = f"Replication factor error: current {replicas}, " \
-                f"target {topic_details[idx]['replicas']}"
-            elif int(partitions) != topic_details[idx]['partitions']:
-                # Partitions number is wrong
-                err = f"Partitions number error: current {partitions}, " \
-                f"target {topic_details[idx]['partitions']}"
-            else:
-                # Topic passed the check
-                topics_ok.append(topic_details[idx]['name'])
-                continue
-            # one of errors happened, add topic to failed list
-            # And add error message
-            failed_topic = (topic_details[idx], err)
-            self.logger.warning(f"Topic {topic_details[idx]['name']}, {err}")
-            topics_failed.append(failed_topic)
-
-        # Check that no errors happened
-        failed = len(topics_failed)
-        assert failed < 1, f"{failed} RP topics has errors"
-
-        topics_available = []
-        topics_unavailable = []
-        # Validate our list against RP
-        for topic in topic_details:
-            # make sure topic is present in RP
-            if topic['name'] not in topics_ok:
-                topics_unavailable.append(topic['name'])
-            else:
-                topics_available.append(topic['name'])
-        unavailable = len(topics_unavailable)
-        assert unavailable < 1, f"{unavailable} topics are missing in RP"
-
-        # Free node that used to create topics
-        self.cluster.free_single(node)
-
-        # Move on to traffic checks
-        topics_to_go = topics_available
-        # Messages to produce
-        message_count = 100
-        self.logger.info(
-            f"Starting Produce/Consume stage for {len(topics_to_go)} topics")
-        producer_errors = []
-        while len(topics_to_go) > 0:
-            topics_to_go, errors = self._write_and_random_read_many_topics(
-                message_count, profile.batch_size, topics_to_go)
-            producer_errors += errors
-            self.logger.info(
-                f"iteration complete, topics left {len(topics_to_go)}")
-
-        total_errors = len(producer_errors)
-        if total_errors > 0:
-            _errors_str = '\n'.join(producer_errors)
-            self.logger.error(f"Producer errors:\n{_errors_str}")
-        assert total_errors < 1, \
-            f"{total_errors} errors detected while sending messages"
-        self.logger.info("Produce/Consume stage complete")
-
-        return
-
-    def _run_producers_with_constant_rate(self, profile, node_topic_count,
-                                          topics):
-        swarm_node_producers = []
-        for topic in topics:
-            swarm_producer = ProducerSwarm(
-                self.test_context,
-                self.redpanda,
-                topic,
-                node_topic_count,
-                profile.message_count,
-                unique_topics=True,
-                messages_per_second_per_producer=profile.
-                messages_per_second_per_producer)
-            swarm_node_producers.append(swarm_producer)
-
-        # Run topic swarm for each topic group
-        for swarm_client in swarm_node_producers:
-            self.logger.info(f"Starting swarm client on node {swarm_client}")
-            swarm_client.start()
-
-        return swarm_node_producers
-
-    def _run_consumers_with_constant_rate(self, profile, node_topic_count,
-                                          topics):
-        swarm_node_consumers = []
-        for topic in topics:
-            swarm_producer = ConsumerSwarm(
-                self.test_context,
-                self.redpanda,
-                topic,
-                node_topic_count,
-                profile.message_count,
-                unique_topics=True,
-                messages_per_second_per_producer=profile.
-                messages_per_second_per_producer)
-            swarm_node_consumers.append(swarm_producer)
-
-        # Run topic swarm for each topic group
-        for swarm_client in swarm_node_consumers:
-            self.logger.info(f"Starting swarm client on node {swarm_client}")
-            swarm_client.start()
-
-        return swarm_node_consumers
-
-    @cluster(num_nodes=12, log_allow_list=RESTART_LOG_ALLOW_LIST)
-    def test_many_topics_throughput(self):
-        """Test creates 11950 topics, and uses client-swarm to
-        generate 1000 messages with 5 msg/sec rate to each topic
-        and validates high watermark values
-
-        Returns:
-            None
-        """
-
-        # Max number of partitions for i3en.xlarge
-        # default settings is 1000 partitions per shard/cpu
-        # (9 nodes × 4 vcpus/shards × 1000) / 3 replicas
-        # 12000
-
-        # Notes on messages params and BW calculations
-        # default msg size is 16KB
-        # 16k * 10 = ~150KB/s per single producer
-        # 10k Producers generate 1.56GB/sec load to the cluster
-        # Rate beyond 40 is unreachable in most cases
-        # Example: 60 msg/sec on 10k topics
-        # [2024-02-26T23:26:50Z INFO  client_swarm] Producer rates: [min=390095, max=682666, avg=506402] bytes/s
-        # => 506402 / 16384 = ~30
-
-        # Get profile for 10k topics
-        tsm = TopicScaleProfileManager()
-        profile = tsm.get_custom_profile("topic_profile_t10k_p1",
-                                         {"batch_size": 1024})
-        # Start kafka
-        self.redpanda.start()
-        # Brokers list suitable for script arguments
-        brokers = ",".join(self.redpanda.brokers_list())
-
-        # Calculate topic counts
-        # This test by default designed for 12 nodes, 9 node cluster, 4 cpus each node
-        # I.e. 12k topics max
-        num_cpus = self.redpanda.get_node_cpu_count()
-        # 1000 partitions per shard
-        max_supported_topics = num_cpus * 1000 * len(self.redpanda.nodes) // 3
-        # Account for system level topics
-        max_supported_topics -= 50
-        # swarm node can generate traffic for 4.5k topics max due to network limitations
-        max_node_topic_count = 4500
-        nodes_available = self.test_context.cluster.available().size()
-        node_topic_count = max_supported_topics // nodes_available
-
-        # swarm node can generate traffic for 4.5k topics max due to network limitations
-        max_node_topic_count = 4500
-        nodes_available = self.test_context.cluster.available().size()
-        node_topic_count = max_supported_topics // nodes_available
-
-        if node_topic_count > max_node_topic_count:
-            self.logger.warning(
-                f"Cluster supports up to {max_supported_topics}"
-                " topics, nodes available for swarm "
-                f"{nodes_available}, topics per swarm node would be "
-                f"{node_topic_count}, which is more than one "
-                f"node can handle ({max_node_topic_count})")
-            node_topic_count = max_node_topic_count
-            self.logger.warning("Setting swarm topic count to MAX of "
-                                f"{max_node_topic_count} per swarm node "
-                                f"({nodes_available} swarm nodes * "
-                                f"{max_node_topic_count} = "
-                                f"{node_topic_count * nodes_available})")
-        else:
-            self.logger.warning("Using swarm topic count of "
-                                f"{node_topic_count} per swarm node "
-                                f"({nodes_available} swarm nodes * "
-                                f"{node_topic_count} = "
-                                f"{node_topic_count * nodes_available})")
-
-        # Grab node to run creation script on it.
-        node = self.cluster.alloc(ClusterSpec.simple_linux(1))[0]
-        topic_prefixes = []
-        # Create unique topics for each swarm node
-        for idx in range(nodes_available):
-            node_topic_name_prefix = f"{profile.topic_name_prefix}{idx}"
-            # Call function to create the topics
-            topic_details = self._create_many_topics(
-                brokers,
-                node,
-                node_topic_name_prefix,
-                node_topic_count,
-                profile.batch_size,
-                profile.num_partitions,
-                profile.num_replicas,
-                profile.use_kafka_batching,
-                skip_name_randomization=True)
-
-            self.logger.info(f"Created {len(topic_details)} topics with "
-                             f"prefix of '{node_topic_name_prefix}'")
-
-            topic_prefixes.append(node_topic_name_prefix)
-        # Free node that used to create topics
-        self.cluster.free_single(node)
-
-        # Do the healthcheck on RP
-        # to make sure that all topics are settle down and have their leader
-
-        # Calculate how much time ideally needed for the producers to finish
-        # Logic is that we sleep for normal running time
-        # And then running checks when swarm nodes running last messages delivery
-        running_time_sec = \
-            profile.message_count // profile.messages_per_second_per_producer
-        self.logger.info(f"Sleeping for {running_time_sec} sec (running time)")
-
-        # Run swarm producers
-        swarm_producers = self._run_producers_with_constant_rate(
-            profile, node_topic_count, topic_prefixes)
-
-        # Run swarm consumers
-        swarm_consumers = self._run_consumers_with_constant_rate(
-            profile, node_topic_count, topic_prefixes)
-
-        # Wait for all messages to be produced
-        # Logic is that we sleep for normal running time
-        # And then running checks when swarm nodes running last messages delivery
-        running_time_sec = \
-            profile.message_count // profile.messages_per_second_per_producer
-        self.logger.info(f"Sleeping for {running_time_sec} sec (running time)")
-        # Just pause for running time to eliminate unnesesary requests
-        # and not put noise into already overloaded network traffic
-        time.sleep(running_time_sec)
-
-        # Run checks if swarm nodes finished
-        self.logger.info("Make sure that swarm node producers are finished")
-        for s in swarm_producers:
-            # account for up to one-third delays
-            s.wait(running_time_sec * 2)
-        self.logger.info("Make sure that swarm node consumers are finished")
-        for s in swarm_consumers:
-            # account for up to one-third delays
-            s.wait(running_time_sec * 2)
-
-        self.logger.info("Calculating high watermarks for all topics")
-
-        # Topic hwm getter
-        def _get_hwm(topic):
-            _hwm = 0
-            for partition in self.rpk.describe_topic(topic):
-                # Add currect high watermark for topic
-                _hwm += partition.high_watermark
-            return _hwm
-
-        # Validate high watermark
-        target_messages_per_node = profile.message_count * node_topic_count
-        hwms = []
-        for topic_prefix in topic_prefixes:
-            # messages per node
-            _topic_names = [
-                f"{topic_prefix}-{idx}" for idx in range(node_topic_count)
-            ]
-            # Use Thread pool to speed things up
-            with concurrent.futures.ThreadPoolExecutor(max_workers=32) as exec:
-                swarmnode_hwms = sum(exec.map(_get_hwm, _topic_names))
-            # save watermark for node
-            hwms.append(swarmnode_hwms)
-
-        assert all([hwm >= target_messages_per_node for hwm in hwms]), \
-            f"Message counts per swarm node mismatch: " \
-            f"target={target_messages_per_node}, " \
-            f"swarm_nodes='''{', '.join([str(num) for num in hwms])}'''"
-
-        return
