+++ tests/rptest/scale_tests/many_topics_test.py
@@ -28,6 +28,8 @@ from rptest.tests.redpanda_test import RedpandaTest
 from rptest.utils.mode_checks import skip_debug_mode
 from rptest.utils.node_operations import NodeOpsExecutor
 from rptest.util import inject_remote_script
+from rptest.services.producer_swarm import ProducerSwarm
+from rptest.services.consumer_swarm import ConsumerSwarm
 from rptest.scale_tests.topic_scale_profiles import TopicScaleProfileManager
 from rptest.clients.python_librdkafka import PythonLibrdkafka
 
@@ -531,3 +533,355 @@ class ManyTopicsTest(RedpandaTest):
 
         self.logger.warn(
             f"Time it took to replace node {unsafe_end - unsafe_start}")
+
+    @cluster(num_nodes=10, log_allow_list=RESTART_LOG_ALLOW_LIST)
+    def test_topic_swarm(self):
+        """Test creates 11950 topics, validates partitions and replicas,
+        produces 100 messages to each topic using batches, consumes
+        messages from random topic in each batch and validates message content
+
+        Returns:
+            None
+        """
+
+        # Max number of partitions for i3en.xlarge
+        # default settings is 1000 partitions per shard/cpu
+        # (9 nodes × 4 vcpus/shards × 1000) / 3 replicas
+        # 12000
+        tsm = TopicScaleProfileManager()
+        profile = tsm.get_profile("topic_profile_t10k_p1")
+
+        # Start kafka
+        self.redpanda.start()
+
+        # Brokers list suitable for script arguments
+        brokers = ",".join(self.redpanda.brokers_list())
+
+        # With current settings, there should be single available node
+        node = self.cluster.alloc(ClusterSpec.simple_linux(1))[0]
+
+        # Call function to create the topics
+        topic_details = self._create_many_topics(
+            brokers,
+            node,
+            profile.topic_name_prefix,
+            profile.topic_count,
+            profile.batch_size,
+            profile.num_partitions,
+            profile.num_replicas,
+            profile.use_kafka_batching,
+            topic_name_length=profile.topic_name_length,
+            skip_name_randomization=False)
+
+        # Validate topics
+        topics = self.rpk.list_topics(detailed=True)
+
+        # Validate topic creation
+        self.logger.info("Validating created topics")
+        topics_ok = []
+        topics_failed = []
+
+        # Searches for a topic in our list
+        def get_topic_index(name):
+            for idx in range(len(topic_details)):
+                if name == topic_details[idx]['name']:
+                    return idx
+            return -1
+
+        # Validate RP topics against requested details
+        for name, partitions, replicas in topics:
+            err = "ok"
+            idx = get_topic_index(name)
+            if idx < 0:
+                # RP has topic that is not ours
+                # Cluster was clean and it is an issue
+                err = f"Unexpected topic found: {name}"
+            elif int(replicas) != topic_details[idx]['replicas']:
+                # Replication factor is wrong
+                err = f"Replication factor error: current {replicas}, " \
+                f"target {topic_details[idx]['replicas']}"
+            elif int(partitions) != topic_details[idx]['partitions']:
+                # Partitions number is wrong
+                err = f"Partitions number error: current {partitions}, " \
+                f"target {topic_details[idx]['partitions']}"
+            else:
+                # Topic passed the check
+                topics_ok.append(topic_details[idx]['name'])
+                continue
+            # one of errors happened, add topic to failed list
+            # And add error message
+            failed_topic = (topic_details[idx], err)
+            self.logger.warning(f"Topic {topic_details[idx]['name']}, {err}")
+            topics_failed.append(failed_topic)
+
+        # Check that no errors happened
+        failed = len(topics_failed)
+        assert failed < 1, f"{failed} RP topics has errors"
+
+        topics_available = []
+        topics_unavailable = []
+        # Validate our list against RP
+        for topic in topic_details:
+            # make sure topic is present in RP
+            if topic['name'] not in topics_ok:
+                topics_unavailable.append(topic['name'])
+            else:
+                topics_available.append(topic['name'])
+        unavailable = len(topics_unavailable)
+        assert unavailable < 1, f"{unavailable} topics are missing in RP"
+
+        # Free node that used to create topics
+        self.cluster.free_single(node)
+
+        # Move on to traffic checks
+        topics_to_go = topics_available
+        # Messages to produce
+        message_count = 100
+        self.logger.info(
+            f"Starting Produce/Consume stage for {len(topics_to_go)} topics")
+        producer_errors = []
+        while len(topics_to_go) > 0:
+            topics_to_go, errors = self._write_and_random_read_many_topics(
+                message_count, profile.batch_size, topics_to_go)
+            producer_errors += errors
+            self.logger.info(
+                f"iteration complete, topics left {len(topics_to_go)}")
+
+        total_errors = len(producer_errors)
+        if total_errors > 0:
+            _errors_str = '\n'.join(producer_errors)
+            self.logger.error(f"Producer errors:\n{_errors_str}")
+        assert total_errors < 1, \
+            f"{total_errors} errors detected while sending messages"
+        self.logger.info("Produce/Consume stage complete")
+
+        return
+
+    def _run_producers_with_constant_rate(self, profile, node_topic_count,
+                                          topics):
+        swarm_node_producers = []
+        for topic in topics:
+            swarm_producer = ProducerSwarm(
+                self.test_context,
+                self.redpanda,
+                topic,
+                node_topic_count,
+                profile.message_count,
+                unique_topics=True,
+                messages_per_second_per_producer=profile.
+                messages_per_second_per_producer)
+            swarm_node_producers.append(swarm_producer)
+
+        # Run topic swarm for each topic group
+        for swarm_client in swarm_node_producers:
+            self.logger.info(f"Starting swarm client on node {swarm_client}")
+            swarm_client.start()
+
+        return swarm_node_producers
+
+    def _run_consumers_with_constant_rate(self, profile, node_topic_count,
+                                          topics, group):
+        swarm_node_consumers = []
+        for topic in topics:
+            swarm_producer = ConsumerSwarm(self.test_context,
+                                           self.redpanda,
+                                           topic,
+                                           group,
+                                           node_topic_count,
+                                           profile.message_count,
+                                           unique_topics=True)
+            swarm_node_consumers.append(swarm_producer)
+
+        # Run topic swarm for each topic group
+        for swarm_client in swarm_node_consumers:
+            self.logger.info(f"Starting swarm client on node {swarm_client}")
+            swarm_client.start()
+
+        return swarm_node_consumers
+
+    @cluster(num_nodes=12, log_allow_list=RESTART_LOG_ALLOW_LIST)
+    def test_many_topics_throughput(self):
+        """Test creates 11950 topics, and uses client-swarm to
+        generate 1000 messages with 5 msg/sec rate to each topic
+        and validates high watermark values
+
+        Returns:
+            None
+        """
+
+        # Max number of partitions for i3en.xlarge
+        # default settings is 1000 partitions per shard/cpu
+        # (9 nodes × 4 vcpus/shards × 1000) / 3 replicas
+        # 12000
+
+        # Notes on messages params and BW calculations
+        # default msg size is 16KB
+        # 16k * 10 = ~150KB/s per single producer
+        # 10k Producers generate 1.56GB/sec load to the cluster
+        # Rate beyond 40 is unreachable in most cases
+        # Example: 60 msg/sec on 10k topics
+        # [2024-02-26T23:26:50Z INFO  client_swarm] Producer rates: [min=390095, max=682666, avg=506402] bytes/s
+        # => 506402 / 16384 = ~30
+
+        def _adjust_topic_count(node_topic_count, nodes_available):
+            # swarm node can generate traffic for 4.5k topics max due to network limitations
+            max_node_topic_count = 4500
+            new_node_topic_count = node_topic_count
+            if new_node_topic_count > max_node_topic_count:
+                self.logger.warning(
+                    f"Cluster supports up to {max_supported_topics}"
+                    " topics, nodes available for swarm "
+                    f"{nodes_available}, topics per swarm node would be "
+                    f"{node_topic_count}, which is more than one "
+                    f"node can handle ({max_node_topic_count})")
+                new_node_topic_count = max_node_topic_count
+                self.logger.warning("Setting swarm topic count to MAX of "
+                                    f"{max_node_topic_count} per swarm node "
+                                    f"({nodes_available} swarm nodes * "
+                                    f"{max_node_topic_count} = "
+                                    f"{node_topic_count * nodes_available})")
+            else:
+                self.logger.warning("Using swarm topic count of "
+                                    f"{node_topic_count} per swarm node "
+                                    f"({nodes_available} swarm nodes * "
+                                    f"{node_topic_count} = "
+                                    f"{node_topic_count * nodes_available})")
+
+            return new_node_topic_count
+
+        # Get profile for 6k topics
+        # This should be enough for default tests with 12 nodes cluster
+        tsm = TopicScaleProfileManager()
+        profile = tsm.get_custom_profile("default", {"batch_size": 1024})
+
+        # Start kafka
+        self.redpanda.start()
+        # Brokers list suitable for script arguments
+        brokers = ",".join(self.redpanda.brokers_list())
+
+        # Calculate topic counts
+        # This test by default designed for 12 nodes, 9 node cluster, 4 cpus each node
+        # I.e. 12k topics max
+        num_cpus = self.redpanda.get_node_cpu_count()
+        # 1000 partitions per shard
+        max_supported_topics = num_cpus * 1000 * len(self.redpanda.nodes) // 3
+        # Account for system level topics
+        max_supported_topics -= 50
+        # Divide available nodes between producers and consumers
+        nodes_available = self.test_context.cluster.available().size()
+
+        produce_nodes = nodes_available // 2
+        consume_nodes = nodes_available // 2
+
+        # uncomment for manual max topics
+        # node_topic_count = max_supported_topics // nodes_available
+
+        # get the target topic counts
+        # It is understood that these numbers will be the same
+        # But for code readability, it is divided into producers and consumers
+        produce_node_topic_count = profile.topic_count // produce_nodes
+        consume_node_topic_count = profile.topic_count // consume_nodes
+
+        self.logger.warning("Checking produce swarm nodes topic count")
+        produce_node_topic_count = _adjust_topic_count(
+            produce_node_topic_count, produce_nodes)
+        self.logger.warning("Checking consume swarm nodes topic count")
+        consume_node_topic_count = _adjust_topic_count(
+            consume_node_topic_count, consume_nodes)
+
+        # Grab node to run creation script on it.
+        node = self.cluster.alloc(ClusterSpec.simple_linux(1))[0]
+        topic_prefixes = []
+        # Create unique topics for each swarm node
+        for idx in range(produce_nodes):
+            node_topic_name_prefix = f"{profile.topic_name_prefix}-{idx}"
+            # Call function to create the topics
+            topic_details = self._create_many_topics(
+                brokers,
+                node,
+                node_topic_name_prefix,
+                produce_node_topic_count,
+                profile.batch_size,
+                profile.num_partitions,
+                profile.num_replicas,
+                profile.use_kafka_batching,
+                topic_name_length=profile.topic_name_length,
+                skip_name_randomization=True)
+
+            self.logger.info(f"Created {len(topic_details)} topics with "
+                             f"prefix of '{node_topic_name_prefix}'")
+
+            topic_prefixes.append(node_topic_name_prefix)
+        # Free node that used to create topics
+        self.cluster.free_single(node)
+
+        # Do the healthcheck on RP
+        # to make sure that all topics are settle down and have their leader
+
+        # Calculate how much time ideally needed for the producers to finish
+        # Logic is that we sleep for normal running time
+        # And then running checks when swarm nodes running last messages delivery
+        running_time_sec = \
+            profile.message_count // profile.messages_per_second_per_producer
+        self.logger.info(f"Sleeping for {running_time_sec} sec (running time)")
+
+        # Run swarm producers
+        swarm_producers = self._run_producers_with_constant_rate(
+            profile, produce_node_topic_count, topic_prefixes)
+
+        # Run swarm consumers
+        _group = "topic_swarm_group"
+        swarm_consumers = self._run_consumers_with_constant_rate(
+            profile, consume_node_topic_count, topic_prefixes, _group)
+
+        # Wait for all messages to be produced
+        # Logic is that we sleep for normal running time
+        # And then running checks when swarm nodes running last messages delivery
+        running_time_sec = \
+            profile.message_count // profile.messages_per_second_per_producer
+        self.logger.info(f"Sleeping for {running_time_sec} sec (running time)")
+        # Just pause for running time to eliminate unnesesary requests
+        # and not put noise into already overloaded network traffic
+        time.sleep(running_time_sec)
+
+        # Run checks if swarm nodes finished
+        self.logger.info("Make sure that swarm node producers are finished")
+        for s in swarm_producers:
+            # account for up to one-third delays
+            s.wait(running_time_sec * 2)
+        self.logger.info("Make sure that swarm node consumers are finished")
+        for s in swarm_consumers:
+            # account for up to one-third delays
+            s.wait(running_time_sec * 2)
+
+        self.logger.info("Calculating high watermarks for all topics")
+
+        # Topic hwm getter
+        def _get_hwm(topic):
+            _hwm = 0
+            for partition in self.rpk.describe_topic(topic):
+                # Add currect high watermark for topic
+                _hwm += partition.high_watermark
+            return _hwm
+
+        # Validate high watermark
+        target_messages_per_node = profile.message_count * produce_node_topic_count
+        hwms = []
+        for topic_prefix in topic_prefixes:
+            # messages per node
+            _topic_names = [
+                f"{topic_prefix}-{idx}"
+                for idx in range(produce_node_topic_count)
+            ]
+            # Use Thread pool to speed things up
+            with concurrent.futures.ThreadPoolExecutor(max_workers=32) as exec:
+                swarmnode_hwms = sum(exec.map(_get_hwm, _topic_names))
+            # save watermark for node
+            hwms.append(swarmnode_hwms)
+
+        assert all([hwm >= target_messages_per_node for hwm in hwms]), \
+            f"Message counts per swarm node mismatch: " \
+            f"target={target_messages_per_node}, " \
+            f"swarm_nodes='''{', '.join([str(num) for num in hwms])}'''"
+
+        return
